+++ shvedova_va/task5/Client.h
@@ -5,16 +5,16 @@ class Client
 {
 	string CardNumber;
 	string FIO;
-	size_t balance;
+	int balance;
 	string PIN;
 	bool status = 1;
 public:
 	Client();
-	Client(const string cn, const string fio, size_t b, const string p, bool s);
+	Client(const string cn, const string fio, int b, const string p, bool s);
 	Client(const Client& c);
 	const string GetCardNumber();
 	const string GetFIO();
-	size_t GetBalance();
+	int GetBalance();
 	const string GetPIN();
 	bool GetStatus();
 	void ChangeBalance(int nb, int side);
